RDMA: Support more than 255 rdma ports
Current code uses many different types when dealing with a port of a RDMA
device: u8, unsigned int and u32. Switch to u32 to clean up the logic.

This allows us to make (at least) the core view consistent and use the
same type. Unfortunately not all places can be converted. Many uverbs
functions expect port to be u8 so keep those places in order not to break
UAPIs.  HW/Spec defined values must also not be changed.

With the switch to u32 we now can support devices with more than 255
ports. U32_MAX is reserved to make control logic a bit easier to deal
with. As a device with U32_MAX ports probably isn't going to happen any
time soon this seems like a non issue.

When a device with more than 255 ports is created uverbs will report the
RDMA device as having 255 ports as this is the max currently supported.

The verbs interface is not changed yet because the IBTA spec limits the
port size in too many places to be u8 and all applications that relies in
verbs won't be able to cope with this change. At this stage, we are
extending the interfaces that are using vendor channel solely

Once the limitation is lifted mlx5 in switchdev mode will be able to have
thousands of SFs created by the device. As the only instance of an RDMA
device that reports more than 255 ports will be a representor device and
it exposes itself as a RAW Ethernet only device CM/MAD/IPoIB and other
ULPs aren't effected by this change and their sysfs/interfaces that are
exposes to userspace can remain unchanged.

While here cleanup some alignment issues and remove unneeded sanity
checks (mainly in rdmavt),

Link: https://lore.kernel.org/r/20210301070420.439400-1-leon@kernel.org

Orabug: 32176425

(cherry picked from commit 1fb7f8973f51ca1a7ffe61a2c779ed15f57f3d82)
cherry-pick-repo=kernel/git/torvalds/linux.git

Conflicts:
	drivers/infiniband/core/cache.c
	drivers/infiniband/core/cm.c
	drivers/infiniband/core/cma.c
	drivers/infiniband/core/cma_configfs.c
	drivers/infiniband/core/counters.c
	drivers/infiniband/core/sa_query.c
	drivers/infiniband/core/verbs.c
	drivers/infiniband/hw/cxgb4/provider.c
	drivers/infiniband/hw/hfi1/ipoib.h
	drivers/infiniband/hw/hfi1/ipoib_main.c
	drivers/infiniband/hw/hns/hns_roce_device.h
	drivers/infiniband/hw/hns/hns_roce_hw_v1.c
	drivers/infiniband/hw/hns/hns_roce_hw_v2.c
	drivers/infiniband/hw/hns/hns_roce_main.c
	drivers/infiniband/hw/mlx4/mlx4_ib.h
	drivers/infiniband/hw/mlx5/counters.c
	drivers/infiniband/hw/mlx5/counters.h
	drivers/infiniband/hw/mlx5/ib_virt.c
	drivers/infiniband/hw/mlx5/main.c
	drivers/infiniband/hw/mlx5/mlx5_ib.h
	drivers/infiniband/hw/mthca/mthca_dev.h
	drivers/infiniband/hw/mthca/mthca_mad.c
	drivers/infiniband/hw/mthca/mthca_qp.c
	drivers/infiniband/hw/ocrdma/ocrdma_ah.c
	drivers/infiniband/hw/ocrdma/ocrdma_ah.h
	drivers/infiniband/hw/ocrdma/ocrdma_verbs.h
	drivers/infiniband/hw/qedr/verbs.c
	drivers/infiniband/hw/qedr/verbs.h
	drivers/infiniband/sw/siw/siw_verbs.c
	drivers/infiniband/sw/siw/siw_verbs.h
	drivers/infiniband/ulp/ipoib/ipoib_main.c
	include/rdma/ib_verbs.h
	include/rdma/rdma_counter.h
	drivers/infiniband/core/cache.c

Contextual conflicts due to code differences between uek6 and upstream.

+++ drivers/infiniband/hw/hns/hns_roce_hw_v1.c
@@ -713,7 +713,7 @@ static int hns_roce_v1_rsv_lp_qp(struct hns_roce_dev *hr_dev)
 	int i, j;
 	u8 queue_en[HNS_ROCE_V1_RESV_QP] = { 0 };
 	u8 phy_port;
-	u8 port = 0;
+	u32 port = 0;
 	u8 sl;
 
 	priv = (struct hns_roce_v1_priv *)hr_dev->priv;
@@ -1735,7 +1735,7 @@ static int hns_roce_v1_chk_mbox(struct hns_roce_dev *hr_dev,
 	return 0;
 }
 
-static int hns_roce_v1_set_gid(struct hns_roce_dev *hr_dev, u8 port,
+static int hns_roce_v1_set_gid(struct hns_roce_dev *hr_dev, u32 port,
 			       int gid_index, const union ib_gid *gid,
 			       const struct ib_gid_attr *attr)
 {
@@ -2724,8 +2724,8 @@ static int hns_roce_v1_m_qp(struct ib_qp *ibqp, const struct ib_qp_attr *attr,
 	u64 *mtts_2 = NULL;
 	int ret = -EINVAL;
 	u64 *mtts = NULL;
-	int port;
-	u8 port_num;
+	u32 port;
+	u32 port_num;
 	u8 *dmac;
 	u8 *smac;
